+++ transformers/examples/sentence_retrieval/run_sentence_retrieval_synthetic.py
@@ -22,6 +22,7 @@ import os
 import sys
 from dataclasses import dataclass, field
 from typing import Optional
+from copy import deepcopy
 
 from datasets import load_dataset
 
@@ -33,6 +34,7 @@ from transformers import (
     AutoModelForMaskedLM,
     AutoTokenizer,
     DataCollatorForLanguageModeling,
+    default_data_collator,
     HfArgumentParser,
     Trainer,
     TrainerWordModifications,
@@ -44,6 +46,7 @@ from transformers.trainer_utils import is_main_process
 # Synthetic languages
 from transformers import modify_inputs_synthetic
 from transformers.synthetic_utils import modify_config
+from utils import modify_config_sentence_retrieval, get_embeddings_word_modif, evaluate_embeddings
 
 logger = logging.getLogger(__name__)
 MODEL_CONFIG_CLASSES = list(MODEL_FOR_MASKED_LM_MAPPING.keys())
@@ -195,7 +198,7 @@ class DataTrainingArguments:
         metadata={
             "help": "File which contains indices in the vocabulary to ignore."
         },
-    )    
+    )
     shift_special: bool = field(
         default=False,
         metadata={
@@ -204,12 +207,21 @@ class DataTrainingArguments:
     )
     # Word modification or syntax modifications
     # (If word modification, then monolingual corpus, if syntax modification, then parallel corpus)
-    syntax_modification: bool = field(
+    bilingual: bool = field(
         default=False,
         metadata={
             "help": "Is the evluation for syntax modification or word modification."
         },
     )
+    pool_type: str = field(
+        default='cls',
+        metadata={
+            "help": "cls/final/middle/higher.\
+                    final = average of last layer \
+                    middle = average of layer n/2 \
+                    higher = average of layer n/2 + 1"
+        },
+    )    
 
     def __post_init__(self):
         if self.dataset_name is None and self.train_file is None and self.validation_file is None:
@@ -358,130 +370,101 @@ def main():
 
     if data_args.line_by_line:
         # When using line_by_line, we just tokenize each nonempty line.
-        padding = "max_length" if data_args.pad_to_max_length else False
+        raise('This argument is hard')
+    else:
+        if not data_args.bilingual:
+            # Otherwise, we tokenize every text, then concatenate them together before splitting them in smaller parts.
+            # We use `return_special_tokens_mask=True` because DataCollatorForLanguageModeling (see below) is more
+            # efficient when it receives the `special_tokens_mask`.
+            def tokenize_function(examples):
+                return tokenizer(examples[text_column_name], return_special_tokens_mask=True)
 
-        def tokenize_function(examples):
-            # Remove empty lines
-            examples["text"] = [line for line in examples["text"] if len(line) > 0 and not line.isspace()]
-            return tokenizer(
-                examples["text"],
-                padding=padding,
-                truncation=True,
-                max_length=data_args.max_seq_length,
-                # We use this option because DataCollatorForLanguageModeling (see below) is more efficient when it
-                # receives the `special_tokens_mask`.
-                return_special_tokens_mask=True,
+            tokenized_datasets = datasets.map(
+                tokenize_function,
+                batched=True,
+                num_proc=data_args.preprocessing_num_workers,
+                remove_columns=column_names,
+                load_from_cache_file=not data_args.overwrite_cache,
             )
 
-        tokenized_datasets = datasets.map(
-            tokenize_function,
-            batched=True,
-            num_proc=data_args.preprocessing_num_workers,
-            remove_columns=[text_column_name],
-            load_from_cache_file=not data_args.overwrite_cache,
-        )
-    else:
-        # Otherwise, we tokenize every text, then concatenate them together before splitting them in smaller parts.
-        # We use `return_special_tokens_mask=True` because DataCollatorForLanguageModeling (see below) is more
-        # efficient when it receives the `special_tokens_mask`.
-        def tokenize_function(examples):
-            return tokenizer(examples[text_column_name], return_special_tokens_mask=True)
-
-        tokenized_datasets = datasets.map(
-            tokenize_function,
-            batched=True,
-            num_proc=data_args.preprocessing_num_workers,
-            remove_columns=column_names,
-            load_from_cache_file=not data_args.overwrite_cache,
-        )
-
-        if data_args.max_seq_length is None:
-            max_seq_length = tokenizer.model_max_length
+            if data_args.max_seq_length is None:
+                max_seq_length = tokenizer.model_max_length
+            else:
+                if data_args.max_seq_length > tokenizer.model_max_length:
+                    logger.warn(
+                        f"The max_seq_length passed ({data_args.max_seq_length}) is larger than the maximum length for the"
+                        f"model ({tokenizer.model_max_length}). Using max_seq_length={tokenizer.model_max_length}."
+                    )
+                max_seq_length = min(data_args.max_seq_length, tokenizer.model_max_length)
         else:
-            if data_args.max_seq_length > tokenizer.model_max_length:
-                logger.warn(
-                    f"The max_seq_length passed ({data_args.max_seq_length}) is larger than the maximum length for the"
-                    f"model ({tokenizer.model_max_length}). Using max_seq_length={tokenizer.model_max_length}."
-                )
-            max_seq_length = min(data_args.max_seq_length, tokenizer.model_max_length)
+            # Otherwise, we tokenize every text, then concatenate them together before splitting them in smaller parts.
+            # We use `return_special_tokens_mask=True` because DataCollatorForLanguageModeling (see below) is more
+            # efficient when it receives the `special_tokens_mask`.
+            def tokenize_function(examples):
+                return tokenizer(examples[text_column_name], return_special_tokens_mask=True)
 
-        # # Main data processing function that will concatenate all texts from our dataset and generate chunks of
-        # # max_seq_length.
-        # def group_texts(examples):
-        #     # Concatenate all texts.
-        #     concatenated_examples = {k: sum(examples[k], []) for k in examples.keys()}
-        #     total_length = len(concatenated_examples[list(examples.keys())[0]])
-        #     # We drop the small remainder, we could add padding if the model supported it instead of this drop, you can
-        #     # customize this part to your needs.
-        #     total_length = (total_length // max_seq_length) * max_seq_length
-        #     # Split by chunks of max_len.
-        #     result = {
-        #         k: [t[i : i + max_seq_length] for i in range(0, total_length, max_seq_length)]
-        #         for k, t in concatenated_examples.items()
-        #     }
-        #     return result
+            tokenized_datasets_source = datasets.map(
+                tokenize_function,
+                batched=True,
+                num_proc=data_args.preprocessing_num_workers,
+                remove_columns=column_names,
+                load_from_cache_file=not data_args.overwrite_cache,
+            )
 
-        # # Note that with `batched=True`, this map processes 1,000 texts together, so group_texts throws away a
-        # # remainder for each of those groups of 1,000 texts. You can adjust that batch_size here but a higher value
-        # # might be slower to preprocess.
-        # #
-        # # To speed up this part, we use multiprocessing. See the documentation of the map method for more information:
-        # # https://huggingface.co/docs/datasets/package_reference/main_classes.html#datasets.Dataset.map
-        # tokenized_datasets = tokenized_datasets.map(
-        #     group_texts,
-        #     batched=True,
-        #     num_proc=data_args.preprocessing_num_workers,
-        #     load_from_cache_file=not data_args.overwrite_cache,
-        # )
+            text_column_name = column_names[1]
+
+            tokenized_datasets_target = datasets.map(
+                tokenize_function,
+                batched=True,
+                num_proc=data_args.preprocessing_num_workers,
+                remove_columns=column_names,
+                load_from_cache_file=not data_args.overwrite_cache,
+            )            
+
+            if data_args.max_seq_length is None:
+                max_seq_length = tokenizer.model_max_length
+            else:
+                if data_args.max_seq_length > tokenizer.model_max_length:
+                    logger.warn(
+                        f"The max_seq_length passed ({data_args.max_seq_length}) is larger than the maximum length for the"
+                        f"model ({tokenizer.model_max_length}). Using max_seq_length={tokenizer.model_max_length}."
+                    )
+                max_seq_length = min(data_args.max_seq_length, tokenizer.model_max_length)            
+
 
     # Make synthetic language modifications if necessary
-    if not data_args.syntax_modification:
-        tokenized_datasets = modify_inputs_synthetic(data_args, training_args, tokenized_datasets, tokenizer=tokenizer)
-
-    # Data collator
-    # This one will take care of randomly masking the tokens.
-    data_collator = DataCollatorForLanguageModeling(tokenizer=tokenizer, mlm_probability=data_args.mlm_probability)
+    if not data_args.bilingual:
+        source_datasets = deepcopy(tokenized_datasets)
+        target_datasets = modify_inputs_synthetic(data_args, training_args, tokenized_datasets, tokenizer=tokenizer)
+    else:
+        source_datasets = tokenized_datasets_source
+        target_datasets = tokenized_datasets_target
 
     # Initialize our Trainer
     trainer = TrainerWordModifications(
         model=model,
         args=training_args,
         data_args=data_args,
-        train_dataset=tokenized_datasets["train"] if training_args.do_train else None,
-        eval_dataset=tokenized_datasets["validation"] if training_args.do_eval else None,
+        # HACK: Using train_dataset as source and eval_dataset as validation
+        train_dataset=source_datasets["train"],
+        eval_dataset=target_datasets["train"],
         tokenizer=tokenizer,
-        data_collator=data_collator,
-    )
+        # data_collator=data_collator,
+    )        
 
-    # Training
-    if training_args.do_train:
-        model_path = (
-            model_args.model_name_or_path
-            if (model_args.model_name_or_path is not None and os.path.isdir(model_args.model_name_or_path))
-            else None
-        )
-        trainer.train(model_path=model_path)
-        trainer.save_model()  # Saves the tokenizer too for easy upload
+    # Collect embeddings for the source and target language
+    source_embeddings = get_embeddings_word_modif(trainer, data_args, 'source')
+    target_embeddings = get_embeddings_word_modif(trainer, data_args, 'target')
 
-    # Evaluation
-    results = {}
-    if training_args.do_eval:
-        logger.info("*** Evaluate ***")
+    # Log
+    logger.info('Finished extracting embeddings.')
 
-        eval_output = trainer.evaluate()
-
-        perplexity = math.exp(eval_output["eval_loss"])
-        results["perplexity"] = perplexity
-
-        output_eval_file = os.path.join(training_args.output_dir, "eval_results_mlm.txt")
-        if trainer.is_world_process_zero():
-            with open(output_eval_file, "w") as writer:
-                logger.info("***** Eval results *****")
-                for key, value in results.items():
-                    logger.info(f"  {key} = {value}")
-                    writer.write(f"{key} = {value}\n")
-
-    return results
+    # Evaluate using cosine similarity as metric for KNN
+    # The first index in the source dataset should be mapped to the first in target, and so on
+    accuracy = evaluate_embeddings(source_embeddings, target_embeddings)
+    logger.info('********')
+    logger.info('The accuracy is: {}%'.format(accuracy))
+    logger.info('********')
 
 
 def _mp_fn(index):